# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GEST_RV_CHUN",
        "description": "Documentation des endpoints API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        }
    ],
    "paths": {
...
        "/medecins": {
            "get": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Liste des médecins",
                "operationId": "ee8d23e77b5ca96cdfc36fd4b2a325ad",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "specialite",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedecinListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Créer un médecin",
                "operationId": "b296aa807e37b7065dd12d7434c14cff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedecinCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedecinResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medecins/{id}": {
            "get": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Afficher un médecin",
                "operationId": "4c914d34aad685f9e769577319629b23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedecinResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Mettre à jour un médecin",
                "operationId": "6ad1fd53a8e6eedf672d502906374ac0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedecinUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedecinResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Supprimer un médecin",
                "operationId": "223ff3968b637fdd6ae601d254ca8dcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medecins/{id}/statut": {
            "patch": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Changer le statut d'un médecin",
                "operationId": "e2c69f61bafec95479ef4ec9798092c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatutUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedecinResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medecins/{id}/planning": {
            "get": {
                "tags": [
                    "Medecins"
                ],
                "summary": "Planning d'un médecin",
                "operationId": "8342c23be2ec3c5bf639f281463a2091",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlanningResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifie"
                    },
                    "data": {
                        "nullable": true
                    },
                    "errors": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "authentication"
                            },
                            "detail": {
                                "type": "string",
                                "example": "Authentification requise pour acceder a cette ressource."
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "nullable": true
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MedecinResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Détail du médecin"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Medecin"
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MedecinListResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Liste des médecins"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedMedecins"
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PlanningResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Planning du medecin"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PlanningMedecin"
                        }
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MedecinCreateRequest": {
                "required": [
                    "nom",
                    "prenom",
                    "email",
                    "telephone",
                    "specialite",
                    "service_medical_id"
                ],
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Diop"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Awa"
                    },
                    "email": {
                        "type": "string",
                        "example": "awa.diop@example.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "771234567"
                    },
                    "specialite": {
                        "type": "string",
                        "example": "Cardiologie"
                    },
                    "service_medical_id": {
                        "type": "string",
                        "example": "2f2c9c2b-2d2a-4e75-b1a6-1a2f1e2a3b4c"
                    }
                },
                "type": "object"
            },
            "MedecinUpdateRequest": {
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Diop"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Awa"
                    },
                    "email": {
                        "type": "string",
                        "example": "awa.diop@example.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "771234567"
                    },
                    "specialite": {
                        "type": "string",
                        "example": "Cardiologie"
                    },
                    "service_medical_id": {
                        "type": "string",
                        "example": "2f2c9c2b-2d2a-4e75-b1a6-1a2f1e2a3b4c"
                    }
                },
                "type": "object"
            },
...
            "StatutUpdateRequest": {
                "required": [
                    "statut"
                ],
                "properties": {
                    "statut": {
                        "type": "string",
                        "enum": [
                            "ACTIF",
                            "INACTIF",
                            "ENCONGE"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}